Wrapper for tqDist

TQDist() and TQAE() are convenience functions that writes a list of trees to text files that can be processed by the C implementation of tqDist Sand2014Quartet. tqDist is then called, and the temporary file is deleted when analysis is complete.

Usage

TQDist(trees)

TQAE(trees)

Arguments

trees

List of phylogenetic trees, of class list or multiPhylo.

Value

TQDist() returns the quartet distance between each pair of trees.

TQAE() returns the number of resolved quartets in agreement between each pair of trees ("A" in Brodal et al. 2013) and the number of quartets that are unresolved in both trees ("E" in Brodal et al. 2013).

Details

Quartets can be resolved in one of five ways, which Brodal2013;textualQuartet and Holt2014;textualQuartet distinguish using the letters A-E, and Estabrook1985;textualQuartet refer to as:

A

s = resolved the same in both trees;

B

d = resolved differently in both trees;

C

r1 = resolved only in tree 1;

D

r2 = resolved only in tree 2 (the comparison tree);

E

u = unresolved in both trees.

References

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)