Direct entry points to "tqDist" functions

Wrappers for functions in "tqDist", which calculate triplet and quartet distances between pairs of trees.

Usage

QuartetDistance(file1, file2)

QuartetAgreement(file1, file2)

PairsQuartetDistance(file1, file2)

OneToManyQuartetAgreement(file1, file2)

AllPairsQuartetDistance(file)

AllPairsQuartetAgreement(file)

TripletDistance(file1, file2)

PairsTripletDistance(file1, file2)

AllPairsTripletDistance(file)

Arguments

file, file1, file2

Paths to files containing a tree or trees in Newick format, possibly created using TQFile().

Value

...Distance() functions return the distance between the requested trees.

...Agreement() functions return the number of triplets or quartets that are:

A, resolved in the same fashion in both trees;

E, unresolved in both trees.

Comparing a tree against itself yields the totals (A+B+C) and (D+E) referred to by Brodal2013;textualQuartet and Holt2014;textualQuartet.

Functions

QuartetDistance(): Returns the quartet distance between the tree. in file1 and the tree in file2.

QuartetAgreement(): Returns a vector of length two, listing [1] the number of resolved quartets that agree (A); [2] the number of quartets that are unresolved in both trees (E). See Brodal2013;textualQuartet.

PairsQuartetDistance(): Quartet distance between the tree on each line of file1 and the tree on the corresponding line of file2.

OneToManyQuartetAgreement(): Quartet distance between the tree in file1 and the tree on each line of file2.

AllPairsQuartetDistance(): Quartet distance between each tree listed in file and each other tree therein.

AllPairsQuartetAgreement(): Quartet status for each pair of trees in file.

TripletDistance(): Triplet distance between the single tree given in each file.

PairsTripletDistance(): Triplet distance between the tree on each line of file1 and the tree on the corresponding line of file2.

AllPairsTripletDistance(): Triplet distance between each tree listed in file and each other tree therein.

References

See also

Author

Algorithms: Brodal2013;textualQuartet; Holt2014;textualQuartet.

C implementation: Sand2014;textualQuartet; modified for portability by Martin R. Smith.

R interface: Martin R. Smith.