Quartet State(s)

Report the status of the specified quartet(s).

QuartetState(tips, bips, splits = bips)

QuartetStates(splits)

Arguments

tips
A four-element array listing a quartet of tips, either by their number (if class numeric) or their name (if class character).

bips
Depreciated; included for compatibility with v1.0.2 and below.

splits
An object that can be induced to a Splits object using as.Splits.

Value

QuartetState returns 0 if the relationships of the four taxa are not constrained by the provided splits, or the index of the closest relative to tips[1], otherwise.

Details

One of the three possible four-taxon trees will be consistent with any set of splits generated from a fully resolved tree. If the taxa are numbered 1 to 4, this tree can be identified by naming the tip most closely related to taxon 1. If a set of splits is generated from a tree that contains polytomies, it is possible that all three four-taxon trees are consistent with the set of splits

Functions

QuartetStates: A convenience wrapper that lists the status of all possible quartets for a given Splits object.

References

Estabrook GF, McMorris FR, Meacham CA (1985). “Comparison of undirected phylogenetic trees based on subtrees of four evolutionary units.” Systematic Zoology, 34(2), 193--200. doi: 10.2307/2413326 .

See also

Examples

{
  nTip <- 6
  trees <- list(ape::rtree(nTip, tip.label=seq_len(nTip), br=NULL),
                ape::rtree(nTip, tip.label=seq_len(nTip), br=NULL))

  trees[[3]] <- TreeTools::CollapseNode(trees[[2]], 9:10)

  QuartetState(c(1, 3, 4, 6), trees[[2]])
  QuartetState(1:4, trees[[1]]) == QuartetState(1:4, trees[[2]])
  QuartetState(c(1, 3, 4, 6), trees[[3]])

  QuartetStates(trees[[2]])
  QuartetStates(trees[[3]])

}

#>  [1] 0 0 0 2 0 0 2 0 0 4 2 0 0 4 4